chore: clean up warnings
